back/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
bcrypt
python-multipart
openai
stripe
python-dotenv
aiohttp
asyncio
uuid
yookassa
alembic